Structure update: separate layout and containers for each page.
