+++ package.json
@@ -10,6 +10,7 @@
     "react-dom": "^19.1.0",
     "react-router-dom": "6",
     "react-scripts": "5.0.1",
+    "sass": "^1.89.2",
     "web-vitals": "^5.0.3"
   },
   "scripts": {
