--- src/components/MainContent.js
@@ -1,12 +1,0 @@
-import React from 'react'
-
-function MainContent() {
-  return (
-    <div>
-      <h3>Main Content</h3>
-      <p>This is where the main data or components go.</p>
-    </div>
-  )
-}
-
-export default MainContent
