--- src/components/Sidebar.js
@@ -1,12 +1,0 @@
-import React from 'react'
-
-function Sidebar() {
-  return (
-    <div>
-      <h3>Sidebar</h3>
-      <p>Navigation or menu goes here</p>
-    </div>
-  )
-}
-
-export default Sidebar
