+++ src/pages/Profile/Profile_layout.js
@@ -1,5 +1,3 @@
-// PLACEHOLDER edit page layout (TEMPLATE)
-
 import React from 'react'
 import {Container, Row, Col} from 'react-bootstrap'
